上线前最后几次上传了

+++ app/src/main/java/com/jinzht/pro/utils/DataCleanManager.java
@@ -143,8 +143,8 @@ public class DataCleanManager {
     // 获取文件
     //Context.getExternalFilesDir() --> SDCard/Android/data/你的应用的包名/files/ 目录，一般放一些长时间保存的数据
     //Context.getExternalCacheDir() --> SDCard/Android/data/你的应用包名/cache/目录，一般存放临时缓存数据
-    public static long getFolderSize(File file) throws Exception {
-        long size = 0;
+    public static double getFolderSize(File file) throws Exception {
+        double size = 0;
         try {
             File[] fileList = file.listFiles();
             for (int i = 0; i < fileList.length; i++) {
